rename Deck file into deck
