+++ src/CardsAPI/tests/test_deck.py
@@ -1,16 +1,16 @@
-from src.CardsAPI import Deck, card
+from src.CardsAPI import deck, card
 
 
 def test_init_deck():
 
-    deck = Deck.Deck()
+    deck = deck.Deck()
     assert deck.top_card_index == 0
     assert len(deck.cards) == 312
     # Check placement of the red Card
     assert ((3 / 4) * 312) - 30 <= deck.red_card_index <= ((3 / 4) * 312) + 30
 
     # Assert not shuffle
-    deck = Deck.Deck(False)
+    deck = deck.Deck(False)
     assert deck.top_card_index == 0
     assert len(deck.cards) == 312
     assert ((3 / 4) * 312) - 30 <= deck.red_card_index <= ((3 / 4) * 312) + 30
@@ -18,7 +18,7 @@ def test_init_deck():
 
 def test_deal_deck():
 
-    deck = Deck.Deck()
+    deck = deck.Deck()
     card = deck.getCard()
     assert type(card) == card.Card
     assert deck.top_card_index == 1
@@ -26,7 +26,7 @@ def test_deal_deck():
 
 def test_shuffle_deck():
 
-    deck = Deck.Deck()
+    deck = deck.Deck()
     try:
         for i in range(500):
             card = deck.getCard()
@@ -36,7 +36,7 @@ def test_shuffle_deck():
         # This shouldn't work because the Deck is only 312 cards long
         assert False
 
-    deck = Deck.Deck()
+    deck = deck.Deck()
 
     for i in range(deck.red_card_index):
         assert not deck.needs_shuffling
@@ -60,6 +60,6 @@ def test_shuffle_deck():
 
 def test_init_shuffle_deck():
 
-    deck1 = Deck.Deck(False)
-    deck2 = Deck.Deck(False)
+    deck1 = deck.Deck(False)
+    deck2 = deck.Deck(False)
     assert deck1.cards[0] == deck2.cards[0]
